Blog_app/requirements.txt
alembic==1.11.0
asgiref==3.7.2
async-generator==1.10
attrs==22.2.0
Babel==2.12.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.0
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-DebugToolbar==0.13.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Secrets==0.1.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
passlib==1.7.4
PySocks==1.7.1
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
speaklater==1.3
SQLAlchemy==2.0.7
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.3.3
wsproto==1.2.0
WTForms==3.0.1